requirements/base.txt
hiredis==1.0.0
glinda==1.0.1
boto3==1.9.245
tornado==6.0.3
redis==3.3.8
prometheus-client==0.7.1
wrapt==1.11.2
asyncio==3.4.3
psycopg2-binary==2.8.3
alembic==1.2.1
aioredis==1.3.0
environs==6.0.0
